Slider for minimap size V1

+++ src/render_map.c
@@ -6,18 +6,17 @@
 /*   By: njennes <njennes@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/05 13:11:29 by njennes           #+#    #+#             */
-/*   Updated: 2022/05/05 15:40:42 by njennes          ###   ########.fr       */
+/*   Updated: 2022/05/05 15:41:06 by njennes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "core.h"
-#include "error_utils.h"
 
 static void	draw_cells(int cell_size, t_vec2 offset);
 static void	render_mmap_wall(int x, int y, int cell_size);
 static void	render_mmap_empty(int x, int y, int cell_size);
 
-void	render_mmap(int zoom)
+void	render_mmap(float zoom)
 {
 	int			cell_size;
 	t_vec2		player_pos;
